backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
scikit-learn==1.5.0
xgboost==2.1.0
pandas==2.2.0
numpy==1.26.0
joblib==1.4.0
pydantic==2.9.0
python-multipart==0.0.9
sqlalchemy==2.0.30
python-docx==1.2.0
cors-middleware==0.0.1
httpx==0.28.1
python-dotenv==1.1.0
